Ability to post comments on a project
It records the comment in the database, but does not show it yet.

Comments will need to be show intertwined with status changes, so
a new data model will be required. This will be done in the next
commit, so I did not bother with displaying the comments for now