Record and display project activity history
A basic ProjectActivity data model which can be explicitly written to to record
an activity. Having an explicit model to record activity makes the querying
interface for project history very easy i.e. if it had to query multiple tables,
pagination, sorting etc would need to be done in-memory and would loose any
performance benefits they are supposed to introduce in the first place.

It is very basic, but it can be easily expanded if needed:

- i.e. it can accept any type of subject in the future via a polymorphic
  association, then as long as there's a view partial to render that type
  of object, it will render in the view correctly
- the subject is optional, but allows adding a display_text as a fallback

this way if we wanted to make the project status change acrtivity display more
interactive or complex, we could introduce a ProjectStatusChange model that
would carry all display properties and logic in itself in a clean, isolated
way.